Implementación del Repository Cambios en el DI: - Ahora DetailViewModel y HomeViewModel dependen de los casos de uso, no del cliente http. Los cuales son inyectados. - El DI inyecta en los casos de uso el repository del que obtienen los datos. - Ahora el DI no depende del cliente http, sino de repository y los casos de uso, los cuales crea para la inyección. Cambios viewModels: - No se les inyecta más el cliente http, sino sus casos de uso asociados correspondientes. - Se eliminaron de ellos las funciones de bajo nivel para obtener datos (ahora en repository). - Para realizar sus acciones ahora derivan el comportamiento en <caso de uso>.execute(), sin involucrarse en detalles. Implementación del repository: - Se creó la interfaz MoviesRepository.kt para aplicar inversión de dependencias. - En la clase MoviesRepositoryImplementation.kt:  -- Por ahora se implementa totalmente la búsqueda de datos, dependiendo de las fuentes locales y externas directamente.  -- Se tienen todas las funciones para la búsqueda de datos, que solían estar en los viewModels, con algunos cambios para que conformen la interfaz de los useCases.
