+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/di/MoviesDependencyInjector.kt
@@ -2,44 +2,27 @@ package edu.dyds.movies.di
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewmodel.compose.viewModel
+import edu.dyds.movies.data.MoviesRepositoryImplementation
+import edu.dyds.movies.domain.repository.MoviesRepository
+import edu.dyds.movies.domain.usecase.MovieDetailsUseCaseImplementation
+import edu.dyds.movies.domain.usecase.MovieDetailsUseCase
+import edu.dyds.movies.domain.usecase.PopularMoviesUseCase
+import edu.dyds.movies.domain.usecase.PopularMoviesUseCaseImplementation
 import edu.dyds.movies.presentation.home.HomeViewModel
 import edu.dyds.movies.presentation.detail.DetailViewModel
-import io.ktor.client.*
-import io.ktor.client.plugins.*
-import io.ktor.client.plugins.contentnegotiation.*
-import io.ktor.http.*
-import io.ktor.serialization.kotlinx.json.*
-import kotlinx.serialization.json.Json
-
-private const val API_KEY = "d18da1b5da16397619c688b0263cd281"
 
 object MoviesDependencyInjector {
 
-    private val tmdbHttpClient =
-        HttpClient {
-            install(ContentNegotiation) {
-                json(Json {
-                    ignoreUnknownKeys = true
-                })
-            }
-            install(DefaultRequest) {
-                url {
-                    protocol = URLProtocol.HTTPS
-                    host = "api.themoviedb.org"
-                    parameters.append("api_key", API_KEY)
-                }
-            }
-            install(HttpTimeout) {
-                requestTimeoutMillis = 5000
-            }
-        }
+    private val repository: MoviesRepository = MoviesRepositoryImplementation()
+    private val popularMoviesUseCase: PopularMoviesUseCase = PopularMoviesUseCaseImplementation(repository)
+    private val movieDetailsUseCase: MovieDetailsUseCase = MovieDetailsUseCaseImplementation(repository)
 
     @Composable
     fun getDetailViewModel(): DetailViewModel {
-        return viewModel { DetailViewModel(tmdbHttpClient) }
+        return viewModel { DetailViewModel(movieDetailsUseCase) }
     }
     @Composable
     fun getHomeViewModel(): HomeViewModel {
-        return viewModel { HomeViewModel(tmdbHttpClient) }
+        return viewModel { HomeViewModel(popularMoviesUseCase) }
     }
 }
