+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/domain/repository/MoviesRepository.kt
@@ -4,5 +4,5 @@ import edu.dyds.movies.domain.entity.Movie
 
 interface MoviesRepository {
     suspend fun getPopularMovies(): List<Movie>
-    suspend fun getMovieDetails(id: Int): Movie
+    suspend fun getMovieDetails(id: Int): Movie?
 }
